# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: tinycdi Workspace Platform — Public API
  version: "0.1.0"
  summary: |
    Public HTTP API of the tinycdi workspace platform (portal origin).
  description: |
    This API is the only interface end users and the portal talk to. It is served
    on the **portal origin**; interactive desktop traffic runs on a separate
    **session origin** reached only through the launch flow described under
    `POST /v1/workspaces/{workspaceId}/connections`.

    ## Multi-tenancy and authorization

    - Authentication is an OIDC-backed session cookie. The verified principal
      (issuer + subject + tenant) is the sole source of ownership: requests never
      carry an owner field and the server never trusts client-supplied identity.
    - Every resource is scoped to the caller's tenant. Regular users see only
      resources they own; tenant administrators see the tenant scope.
      Cross-tenant access returns `404 NOT_FOUND` (the resource is not
      distinguishable from non-existent), never a leak of existence.
    - Authorization to act on a workspace (view, start, stop, delete, connect,
      attach, purge) requires ownership or tenant-admin role; otherwise
      `403 FORBIDDEN` / `404 NOT_FOUND`.
    - End users never write Kubernetes custom resources. The API translates
      accepted intents into `Workspace` CRs through its own service account via a
      transactional outbox (see ADR 0002); Kubernetes RBAC denies direct CR
...

    ## Idempotency

...

...

...
servers:
...
security:
  - sessionCookie: []
...
paths:
...
  /v1/data:
    get:
      operationId: listRetainedData
      summary: List retained data disks visible to the caller
      description: |
        Returns retained-disk records owned by the caller (tenant
        administrators see the tenant scope). Each record carries a fresh
        `purgeConfirmationNonce` required by `POST /v1/data/{dataId}/purge`.
      tags: [data]
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageToken"
      responses:
        "200":
          description: Page of retained data records.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetainedDataList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/data/{dataId}/attach:
    post:
      operationId: attachRetainedData
      summary: Attach a retained disk to a new workspace
      description: |
        Creates a **new workspace** whose disk is the retained record. The
        record must be in state `Retained` and is claimed exclusively: the
        transition `Retained → Attaching → Attached` is a conditional update in
        the same transaction as the quota reservation and outbox intent, so
        attach and purge mutually exclude each other and a disk has at most one
        consumer at a time. If the new workspace is deleted before the attach
        completes, the disk returns to `Retained`; it is never garbage-collected
        with the old workspace.

        The resulting workspace is a new runtime identity: new
        `runtimeGeneration` and enrollment token, and the old workspace's
        credentials are invalidated on first boot via re-enrollment.

        Only the owner recorded on the retained record (or a tenant admin) may
        attach. Idempotent via `Idempotency-Key`.
      tags: [data]
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/DataId"
        - $ref: "#/components/parameters/IdempotencyKeyRequired"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachDataRequest"
      responses:
        "201":
          description: Workspace created from the retained disk.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceView"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/data/{dataId}/purge:
    post:
      operationId: purgeRetainedData
      summary: Irreversibly destroy a retained disk
      description: |
        Schedules destruction of a retained disk. Requires the
        `confirmationNonce` from the caller's latest read of the record
        (`GET /v1/data`) — the nonce is short-lived, single-use and bound to the
        record, principal and record state, so replaying an old nonce fails with
        `400 INVALID_REQUEST`.

        Only records in state `Retained` can be purged: `Attaching`/`Attached`
        disks return `409 INVALID_STATE` (detach/delete the consuming workspace
        first), and a concurrent attach wins or loses the race atomically —
        never both. Purge is asynchronous; the record reports `Purging` until
        the disk is actually deleted, and storage quota stays reserved until
        then.
      tags: [data]
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/DataId"
        - $ref: "#/components/parameters/IdempotencyKeyOptional"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurgeDataRequest"
      responses:
        "202":
          description: Purge accepted; record is now `Purging`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetainedDataView"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: tcdi_session
      description: |
        OIDC-backed session cookie issued by the portal login flow
        (`Secure`, `HttpOnly`, `SameSite=Strict` on the portal origin).
        Required on every endpoint.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        CSRF token obtained from the session bootstrap payload and echoed on
        every mutating request. Combined with `SameSite=Strict` cookies this
        blocks cross-site request forgery; a missing or mismatched token
        returns `403 CSRF_FAILED`. Required on all non-GET operations.
  parameters:
...
    DataId:
      name: dataId
      in: path
      required: true
      description: Server-generated retained-data record identifier.
      schema:
        type: string
        pattern: "^rd_[A-Za-z0-9]{8,64}$"
        example: rd_01J4Z9W2PFK8G4TQ3M7H1R5N0A
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated unique key (recommended: UUIDv4/ULID). Same key + same
        request body replays the recorded result; same key + different body
        returns `409 IDEMPOTENCY_CONFLICT`. Scoped to the principal; retained
        for 24 h.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        example: "9f3b2c44-3d7a-4b1e-9c2d-1f8a0b6c5d4e"
    IdempotencyKeyOptional:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Optional dedup key honored the same way as on create; the operation is
        naturally idempotent without it.
      schema:
        type: string
        minLength: 8
        maxLength: 128
    PageLimit:
      name: limit
      in: query
      required: false
      description: Maximum items per page (default 50, max 200).
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    PageToken:
      name: pageToken
      in: query
      required: false
      description: Opaque cursor from a previous response's `nextPageToken`.
      schema:
        type: string
  responses:
    BadRequest:
      description: |
        The request is malformed or violates the contract — `INVALID_REQUEST`
        (bad JSON, unknown field, missing confirmation nonce, invalid cursor).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, expired or invalid session — `UNAUTHENTICATED`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: |
        Authenticated but not allowed — `FORBIDDEN` for authorization failures,
        `CSRF_FAILED` for a missing/mismatched CSRF token on a mutation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: |
        Resource does not exist or is not visible to the caller's tenant —
        `NOT_FOUND`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: |
        State or idempotency conflict — `IDEMPOTENCY_CONFLICT` (key reused with
        a different body), `INVALID_STATE` (operation not valid in the current
        phase/record state), `CONNECTION_IN_USE` (a live interactive lease
        exists and `takeover` was not set), or `QUOTA_EXHAUSTED` (tenant/user
        quota has no headroom).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnprocessableEntity:
      description: |
        Semantically invalid reference — `INVALID_TEMPLATE` (templateRef is
        unknown, unpublished, not permitted for this tenant, or incompatible
        with the retained disk's runtime).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Rate limit exceeded — `RATE_LIMITED`. Honor `Retry-After`.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Unexpected server error — `INTERNAL`. Safe to retry.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unavailable:
      description: |
        Transient dependency failure — `UNAVAILABLE` (store/broker
        unreachable). Safe to retry with backoff.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message, retryable, requestId]
      description: Uniform error body returned for every 4xx/5xx response.
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          description: Human-readable detail; not part of the contract.
          example: "workspace ws_01J4Z8KQ2M9XNBV3T7YH0R6D5E is not in a startable phase"
        retryable:
          type: boolean
          description: |
            Whether retrying the same request may succeed without user action
            (see the `ErrorCode` table for per-code semantics).
        requestId:
          type: string
          description: Server-generated request ID for support/log correlation.
          example: req_01J4ZA1X0YJ8H6K0Q3MW1N5B2P
    WorkspacePhase:
      type: string
      description: |
        Lifecycle phase — a summary; `conditions` carry the operational detail.
        `Ready` requires the runtime to be actually connectable (display +
        streaming endpoint checked), not merely scheduled.
      enum: [Pending, Provisioning, Ready, Stopping, Stopped, Failed, Terminating]
    DesiredState:
      type: string
      description: User intent for the runtime.
      enum: [Running, Stopped]
    DataPolicy:
      type: string
      description: |
        `Ephemeral` destroys disk data on stop/delete; `Retain` keeps the
        persistent volume across stop and moves it to the retained inventory on
        delete.
      enum: [Ephemeral, Retain]
    RuntimeKind:
      type: string
      enum: [LinuxContainer, WindowsVM]
    ExperienceKind:
      type: string
      enum: [Desktop, Browser]
...
    ConditionStatus:
      type: string
      enum: ["True", "False", "Unknown"]
    WorkspaceCondition:
      type: object
      additionalProperties: false
      required: [type, status, reason, lastTransitionTime]
      description: |
        Summary of a Kubernetes-style condition projected for UI use. `reason`
        is a machine-readable CamelCase string; `message` is optional detail.
      properties:
        type:
          $ref: "#/components/schemas/ConditionType"
        status:
          $ref: "#/components/schemas/ConditionStatus"
        reason:
          type: string
          example: BootDeadlineExceeded
        message:
          type: string
          description: Optional human-readable detail; not part of the contract.
        lastTransitionTime:
          type: string
          format: date-time
    TemplateSummary:
      type: object
      additionalProperties: false
      required: [id, name, revision, runtime, experience]
      description: Immutable template+revision the workspace was created from.
      properties:
        id:
          type: string
          pattern: "^tpl_[A-Za-z0-9][A-Za-z0-9-]{6,62}[A-Za-z0-9]$"
          example: tpl_01J4ZB3N1RXD7P2V8W5K0H6Q4M
        name:
          type: string
          example: linux-firefox-desktop
        revision:
          type: integer
          description: Immutable template revision number.
          example: 7
        runtime:
          $ref: "#/components/schemas/RuntimeKind"
        experience:
          $ref: "#/components/schemas/ExperienceKind"
    WorkspaceView:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - template
        - phase
        - conditions
        - desiredState
        - dataPolicy
        - createdAt
        - updatedAt
      description: |
        Public view of a workspace. It deliberately omits control-plane
        internals: no runtime credentials (never serialized), no Kubernetes
        names/UIDs, and no `runtimeGeneration`/`runtimeUID` fencing values —
        the UI only needs `phase` + `conditions` to decide connectability.
      properties:
        id:
          type: string
          pattern: "^ws_[A-Za-z0-9]{8,64}$"
          example: ws_01J4Z8KQ2M9XNBV3T7YH0R6D5E
        name:
          type: string
          description: Caller-assigned display name; unique per owner.
          example: research-desktop
        template:
          $ref: "#/components/schemas/TemplateSummary"
        phase:
          $ref: "#/components/schemas/WorkspacePhase"
        conditions:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceCondition"
          description: Current condition summary (may be empty while Pending).
        desiredState:
          $ref: "#/components/schemas/DesiredState"
        dataPolicy:
          $ref: "#/components/schemas/DataPolicy"
        retainedDataRef:
          type: string
          pattern: "^rd_[A-Za-z0-9]{8,64}$"
          description: Present when the workspace consumes a retained disk.
        failureReason:
          type: string
          description: Short machine-readable cause when `phase` is `Failed`.
          example: BootDeadlineExceeded
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    RetainedDataView:
      type: object
      additionalProperties: false
      required:
        - id
        - state
        - sizeGib
        - runtime
        - sourceWorkspaceName
        - retainedAt
        - purgeConfirmationNonce
      properties:
        id:
          type: string
          pattern: "^rd_[A-Za-z0-9]{8,64}$"
        state:
          $ref: "#/components/schemas/RetainedDataState"
        sizeGib:
          type: integer
          minimum: 1
          example: 20
        runtime:
          $ref: "#/components/schemas/RuntimeKind"
          description: Runtime family of the disk, to constrain attach templates.
        sourceWorkspaceName:
          type: string
          description: Display name of the workspace the disk was retained from.
          example: research-desktop
        consumingWorkspaceId:
          type: string
          pattern: "^ws_[A-Za-z0-9]{8,64}$"
          description: Present while `Attaching`/`Attached` — the workspace consuming the disk.
        retainedAt:
          type: string
          format: date-time
          description: When the disk entered the retained inventory.
        purgeConfirmationNonce:
          type: string
          description: |
            Short-lived, single-use nonce bound to this record, the caller's
            principal and the current record state. Echo it in
            `POST /v1/data/{id}/purge` to confirm destruction; a fresh nonce is
            issued on every read.
    RetainedDataList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RetainedDataView"
        nextPageToken:
          type: string
    AttachDataRequest:
      type: object
      additionalProperties: false
      required: [name, templateRef]
      description: |
        Attach intent — equivalent to `POST /v1/workspaces` with
        `retainedDataRef`. `dataPolicy` of the new workspace is `Retain` by
        definition; the template's runtime must match the disk's runtime.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
          pattern: "^[a-z0-9][a-z0-9-]{0,126}[a-z0-9]$|^[a-z0-9]$"
          example: restored-desktop
        templateRef:
          type: string
          pattern: "^tpl_[A-Za-z0-9][A-Za-z0-9-]{6,62}[A-Za-z0-9]$"
        desiredState:
          $ref: "#/components/schemas/DesiredState"
    PurgeDataRequest:
      type: object
      additionalProperties: false
      required: [confirmationNonce]
      properties:
        confirmationNonce:
          type: string
          minLength: 8
          description: Value of `purgeConfirmationNonce` from the latest read of the record.